+++ src/daemon/rest/RestHandler.cpp
@@ -918,7 +918,7 @@ void RestHandler::apiAppOutputView(const HttpRequest &message)
 
 	long pos = getHttpQueryValue(message, HTTP_QUERY_KEY_stdout_position, 0, 0, 0);
 	int index = getHttpQueryValue(message, HTTP_QUERY_KEY_stdout_index, 0, 0, 0);
-	int maxSize = getHttpQueryValue(message, HTTP_QUERY_KEY_stdout_maxsize, APP_STD_OUT_VIEW_DEFAULT_SIZE, 1024, APP_STD_OUT_VIEW_DEFAULT_SIZE);
+	long maxSize = getHttpQueryValue(message, HTTP_QUERY_KEY_stdout_maxsize, APP_STD_OUT_VIEW_DEFAULT_SIZE, 1024, APP_STD_OUT_VIEW_DEFAULT_SIZE);
 	std::string processUuid = getHttpQueryString(message, HTTP_QUERY_KEY_process_uuid);
 
 	checkAppAccessPermission(message, appName, false);
